Move extension CR utils from `common` to `extensions` package
