+++ pkg/operation/botanist/component/extensions/infrastructure/infrastructure_test.go
@@ -23,6 +23,7 @@ import (
 	gardencorev1beta1 "github.com/gardener/gardener/pkg/apis/core/v1beta1"
 	v1beta1constants "github.com/gardener/gardener/pkg/apis/core/v1beta1/constants"
 	extensionsv1alpha1 "github.com/gardener/gardener/pkg/apis/extensions/v1alpha1"
+	"github.com/gardener/gardener/pkg/extensions"
 	"github.com/gardener/gardener/pkg/logger"
 	mockclient "github.com/gardener/gardener/pkg/mock/controller-runtime/client"
 	mocktime "github.com/gardener/gardener/pkg/mock/go/time"
@@ -292,7 +293,7 @@ var _ = Describe("#Interface", func() {
 
 		It("should not return error when it's deleted successfully", func() {
 			defer test.WithVars(
-				&common.TimeNow, mockNow.Do,
+				&extensions.TimeNow, mockNow.Do,
 			)()
 			mockNow.EXPECT().Do().Return(now.UTC()).AnyTimes()
 
@@ -317,7 +318,7 @@ var _ = Describe("#Interface", func() {
 
 		It("should return error when it's not deleted successfully", func() {
 			defer test.WithVars(
-				&common.TimeNow, mockNow.Do,
+				&extensions.TimeNow, mockNow.Do,
 			)()
 			mockNow.EXPECT().Do().Return(now.UTC()).AnyTimes()
 
@@ -438,7 +439,7 @@ var _ = Describe("#Interface", func() {
 		It("should properly restore the infrastructure state if it exists", func() {
 			defer test.WithVars(
 				&infrastructure.TimeNow, mockNow.Do,
-				&common.TimeNow, mockNow.Do,
+				&extensions.TimeNow, mockNow.Do,
 			)()
 			mockNow.EXPECT().Do().Return(now.UTC()).AnyTimes()
 
@@ -468,7 +469,7 @@ var _ = Describe("#Interface", func() {
 	Describe("#Migrate", func() {
 		It("should migrate the resources", func() {
 			defer test.WithVars(
-				&common.TimeNow, mockNow.Do,
+				&extensions.TimeNow, mockNow.Do,
 			)()
 			mockNow.EXPECT().Do().Return(now.UTC()).AnyTimes()
