+++ pkg/operation/botanist/component/extensions/worker/worker.go
@@ -24,9 +24,9 @@ import (
 	v1beta1constants "github.com/gardener/gardener/pkg/apis/core/v1beta1/constants"
 	gardencorev1beta1helper "github.com/gardener/gardener/pkg/apis/core/v1beta1/helper"
 	extensionsv1alpha1 "github.com/gardener/gardener/pkg/apis/extensions/v1alpha1"
+	"github.com/gardener/gardener/pkg/extensions"
 	"github.com/gardener/gardener/pkg/operation/botanist/component"
 	"github.com/gardener/gardener/pkg/operation/botanist/component/extensions/operatingsystemconfig"
-	"github.com/gardener/gardener/pkg/operation/common"
 	"github.com/gardener/gardener/pkg/utils"
 
 	"github.com/Masterminds/semver"
@@ -250,7 +250,7 @@ func (w *worker) deploy(ctx context.Context, operation string) (extensionsv1alph
 
 // Restore uses the seed client and the ShootState to create the Worker resources and restore their state.
 func (w *worker) Restore(ctx context.Context, shootState *gardencorev1alpha1.ShootState) error {
-	return common.RestoreExtensionWithDeployFunction(
+	return extensions.RestoreExtensionWithDeployFunction(
 		ctx,
 		w.client,
 		shootState,
@@ -262,7 +262,7 @@ func (w *worker) Restore(ctx context.Context, shootState *gardencorev1alpha1.Sho
 
 // Migrate migrates the Worker resource.
 func (w *worker) Migrate(ctx context.Context) error {
-	return common.MigrateExtensionCR(
+	return extensions.MigrateExtensionCR(
 		ctx,
 		w.client,
 		func() extensionsv1alpha1.Object { return &extensionsv1alpha1.Worker{} },
@@ -273,7 +273,7 @@ func (w *worker) Migrate(ctx context.Context) error {
 
 // Destroy deletes the Worker resource.
 func (w *worker) Destroy(ctx context.Context) error {
-	return common.DeleteExtensionCR(
+	return extensions.DeleteExtensionCR(
 		ctx,
 		w.client,
 		func() extensionsv1alpha1.Object { return &extensionsv1alpha1.Worker{} },
@@ -284,7 +284,7 @@ func (w *worker) Destroy(ctx context.Context) error {
 
 // Wait waits until the Worker resource is ready.
 func (w *worker) Wait(ctx context.Context) error {
-	return common.WaitUntilExtensionCRReady(
+	return extensions.WaitUntilExtensionCRReady(
 		ctx,
 		w.client,
 		w.logger,
@@ -309,7 +309,7 @@ func (w *worker) Wait(ctx context.Context) error {
 
 // WaitMigrate waits until the Worker resources are migrated successfully.
 func (w *worker) WaitMigrate(ctx context.Context) error {
-	return common.WaitUntilExtensionCRMigrated(
+	return extensions.WaitUntilExtensionCRMigrated(
 		ctx,
 		w.client,
 		func() extensionsv1alpha1.Object { return &extensionsv1alpha1.Worker{} },
@@ -322,7 +322,7 @@ func (w *worker) WaitMigrate(ctx context.Context) error {
 
 // WaitCleanup waits until the Worker resource is deleted.
 func (w *worker) WaitCleanup(ctx context.Context) error {
-	return common.WaitUntilExtensionCRDeleted(
+	return extensions.WaitUntilExtensionCRDeleted(
 		ctx,
 		w.client,
 		w.logger,
